backend/requirements.txt
flask
flask-cors
requests
gunicorn
base58
APScheduler
python-telegram-bot
SQLAlchemy
psycopg2-binary
python-dotenv 